Add printing separator before output
The optional --separator command line option causes to write out the
specified sparator on a new line to the output stream before writing
out the remote data.  Useful when printing to stdout.

The separator string can contain the standard escape charaters (\n,
\a, \b, \f, \t, \v, \\) (see man printf for their meaning) and also
the following replacement sequences:

    %u: will be replaced with the URL of the remote file
    %%: will be replaced with a literal percentage character (%)

Unknown backslash escapes will yield the character following the
backslash with the preceding backslash omitted (e.g. '\!' will produce
a plain '!' in the outout.)  On the other hand, an unknown replacement
sequence will be produced to the output as is (e.g. '%H' will produce
'%H' in the output.)